requirements.txt
asgiref==3.8.1
cffi==1.17.1
cryptography==44.0.2
Django==5.1.7
djangorestframework==3.15.2
djangorestframework_simplejwt==5.5.0
mysqlclient==2.2.7
pycparser==2.22
PyJWT==2.9.0
sqlparse==0.5.3
tzdata==2025.2
